+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixes
 - Increase max result limit of data sources for target groups (200) and IP blocks (1000), as a workaround for pagination issues.
 - Add email filter for user data source, fixes pagination issues for users.
+- Fix name validation for Dataplatform resources
 
 ## 6.4.14
 ### Features